% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/sample_func.R
\name{s_func}
\alias{s_func}
\title{Create sampling function
This is a function to create a sample of the length n for a given pdf}
\usage{
s_func(n, pdf, a, b, C)
}
\arguments{
\item{n}{is the number elements to be returned buy the function}

\item{pdf}{is the user give probability density function used to create the sample}

\item{a}{is the lower bound of the value of elements is the sample}

\item{b}{is the upper bound of the value of elements is the sample}

\item{C}{is >= to the max value of the pdf}
}
\value{
a vector of n sample that have been excepted as being valid according to pdf of the
interval (a, b)
}
\description{
Create sampling function
This is a function to create a sample of the length n for a given pdf
}
\examples{
exppdf <- function(x){
exp(x)/(1+exp(x))^2}
s_func(n = 1000, pdf = exppdf, a = -4, b = 4, C= 1)

normpdf <- function(x){
dnorm(x)}
s_func(n = 1000, pdf = normpdf, a = -4, b = 4, C= 1)
}
